==> experiments/liftedindex_lr/01_download_docs.ipynb <==
# Download Docs
This notebook downloads documentation from the IGRA2 website.

- igra2-station-list.txt - The data in this file allows us to identify which station ids we are interested in
- igra2-list-format.txt - Documents the data structure of igra2-station-list.txt
- igra2-data-format.txt - Documents the data structure of the data files

Update the following parameters in the first cell to accomodate your installation:

- DST_PATH - The location to download the files into

import os
import requests

SRC_PATH = 'https://www.ncei.noaa.gov/data/integrated-global-radiosonde-archive/doc'
DST_PATH = '/usr/datalake/bronze/igra/doc'
FILES = ['igra2-station-list.txt', 'igra2-data-format.txt', 'igra2-list-format.txt']

# Make sure the destination path exists
os.makedirs(DST_PATH, exist_ok=True)

# Loop through the list of files
for filename in FILES:
    src = f'{SRC_PATH}/{filename}'
    dst = f'{DST_PATH}/{filename}'

    # If the file already exists, don't download it again
    if os.path.exists(dst):
        print(f"File exists: {filename}")
        continue

    # Download the file
    with requests.get(src) as r:
        open(dst, 'wb').write(r.content)

    print(f"Downloaded: {filename}")
