==> experiments/liftedindex_lr/02_transform_station_list.ipynb <==
# Transform Station List
The raw station list is in an inconvenient format. This notebook uses Pandas to transform it into a format that will allow for the structure to be inferred, reducing our workload.

Update the following parameters in the first cell to accomodate your installation:

- BRONZE_STATION_LIST_PATH - The location of the raw igra2-station-list.txt file
- SILVER_STATION_LIST_PATH - The location to save the CSV version of the file

import os
import pandas as pd

BRONZE_STATION_LIST_PATH = '/usr/datalake/bronze/igra/doc/igra2-station-list.txt'
SILVER_STATION_LIST_PATH = '/usr/datalake/silver/igra/doc/igra2-station-list.csv'

# Get the path without the filename
dest_path = '/'.join(SILVER_STATION_LIST_PATH.replace('\\', '/').split('/')[:-1]) 
print(f'Output folder: {dest_path}')

# Make sure the destination path exists
os.makedirs(dest_path, exist_ok=True)

# Variable not needed anymore
del dest_path

# Define the fixed width intervals
colspecs = [
    (0, 11),        # id
    (12, 20),       # latitude
    (21, 30),       # longitude
    (31, 37),       # elevation
    (38, 40),       # state
    (41, 71),       # name
    (72, 76),       # fstyear
    (77, 81),       # lstyear
    (82, 88)        # nobs
]

# There's no header row so we must specify our own column names
names = ['id', 'latitude', 'longitude', 'elevation', 'state', 'name', 'fst_year', 'lst_year', 'nobs']

# read_fwf is the fixed width file reader for Pandas. The colspecs parameter specifies the fixed column
# ranges. The names parameter specifies the column names. Passing header=None tells Pandas the first
# row contains data, not column names.
df = pd.read_fwf(BRONZE_STATION_LIST_PATH, colspecs=colspecs, header=None, names=names, index_col=0)

# Confirm the data types are correct (they are)
df.info()

# View a sampling for sanity checks (it is sane). Many records have a null state field, so NaN is expected.
df.head()

# Confirm that the state field is parsed correctly by viewing a row we know contains a value
df.loc['USW00094730']

# Save the dataframe as a CSV file
df.to_csv(SILVER_STATION_LIST_PATH)
