==> experiments/liftedindex_lr/03_download_data.ipynb <==
# Download Data
This notebook will download the actual IGRA2 data. By default, we download data for those stations in Tornado Alley that are currently collecting data.

## Data-Por
We are obtaining the data-por version of the IGRA2 data because it contains the full history of observations. The more data the better. By default, we download 23 data files totalling about 1.6 GB in size. You do not need to unzip these files as the olieigra reader can process them in their zipped form.

## Parameterization
Update these parameters to suit your needs

- SILVER_STATION_LIST_PATH - Full path to the igra2-station-list.csv file (See README in this folder if you don't have this file)
- DST_PATH - Folder to download the data files into
- QUERY - SQL Statement to select the stations for which to download data

import os
import pandas as pd
import requests
from pandasql import sqldf

SILVER_STATION_LIST_PATH = '/usr/datalake/silver/igra/doc/igra2-station-list.csv'
SRC_PATH = 'https://www.ncei.noaa.gov/data/integrated-global-radiosonde-archive/access/data-por'
DST_PATH = '/usr/datalake/bronze/igra/data-por'

# Select stations in Tornado Alley that are currently collecting data
QUERY = '''SELECT id, state, name, fst_year, nobs
FROM df 
WHERE STATE IN ('CO', 'IA', 'IL', 'KS', 'MN', 'MO', 'ND', 'NE', 'OK', 'SD', 'TX', 'WI')
    AND lst_year = 2025
ORDER BY state, name'''

# Make sure the destination path exists
os.makedirs(DST_PATH, exist_ok=True)

# Load the CSV and data sanity check
df = pd.read_csv(SILVER_STATION_LIST_PATH, index_col=0, dtype={ 'state': str })
df.head()

# Perform the query on the data frame
results = sqldf(QUERY)

# Output which stations we matched
results

for _, station in results.iterrows():
    filename = f"{station.id}-data.txt.zip"
    url = f"{SRC_PATH}/{filename}"
    local = f"{DST_PATH}/{filename}"

    if os.path.exists(local):
        print(f"File already exists: {local}")
        continue

    r = requests.get(url)
    open(local, 'wb').write(r.content)
    print(f"Downloaded {local}")
