==> experiments/liftedindex_lr/11_calculate_li.ipynb <==
# Calculate LI
This notebook calcualtes the lifted index values for us to train models with. See the README to learn what the Lifted Index is. It reads in the raw data-por IGRA data. It then unpivots the data, converting columns back into pressure levels. This data can then be used by MetPy to calculate the Lifted Index using the actual physics algorithms. A CSV file is geratated for each input file.

**This notebook will take a long time to complete!**

Update the following parameters in the first cell to accomodate your installation:

- SILVER_GPH20S10K_PATH - The location to read the transformed CSV files
- SILVER_LI_PATH - Location to save the calculated Lifted Index values

import math
import os
from datetime import datetime
import numpy as np
from metpy.calc import parcel_profile, lifted_index
from metpy.units import units
import olieigra

BRONZE_DATA_POR_PATH = '/usr/datalake/bronze/igra/data-por'
SILVER_LI_PATH = '/usr/datalake/silver/igra/li'

# If you need to install olieigra, uncomment and execute this line. View the README in the project root for instructions
# on how to build or download this file.
#%pip install /lakehouse/default/Files/libs/olieigra-0.0.1-py3-none-any.whl

# Make sure the destination folder exists
os.makedirs(SILVER_LI_PATH, exist_ok=True)

# from datetime import datetime
class LiftedIndex(olieigra.Callbacks):
    """Calculate the lifted index from data-por igra data"""

    def __init__(self, dst_path: str, min_effective_date: datetime):
        super().__init__()
        self.dst_path = dst_path
        self.min_effective_date = min_effective_date
        self.filtered = 0
        self.errors = 0
        self.data = 0
        self.hout = ''
        self.filename = ''
        self.writer = None

    def start_file(self, filename: str) -> bool:
        """Decide if we want to process the file. If so, reset state and start writing to a
        temporary file."""

        # An IGRA2 file should end with -data.txt
        if not filename.endswith('-data.txt'):
            print(f'Skipping {filename}. Not sure what to do with it.')
            return False

        # Set the desired destination filename
        dst_filename = f'{self.dst_path}/{filename}'
        dst_filename = dst_filename.replace("-data.txt", "-data-li.csv")

        # Skip this file if it has already been processed
        if os.path.exists(dst_filename):
            print(f'Skipping {filename}. Destination file already exists.')
            return False

        # If we got here, we are going to process the file
        print(f'Processing {filename}.')

        # Write to a temp file
        self.filename = dst_filename.replace('-data-li.csv', '-data-li.partial.csv')
        self.writer = open(self.filename, 'w', encoding='UTF-8')

        # Reset the filtered record count
        self.filtered = 0
        self.errors = 0
        self.data = 0
        self.hout = ''

        # Write the header row
        self.writer.write('id,effective_date,hour,li\n')

        # Tell olieigra to continue processing
        return True

    def finish_file(self, headers: int, rows: int):
        """File processing is complete. Clean up and provide user feedback."""

        # Close the temporary file
        self.writer.close()

        # Rename the temporary file
        dst_renamed = self.filename.replace('.partial.csv', '.csv')
        os.rename(self.filename, dst_renamed)

        # Calculate the number of records written
        loaded = headers - self.filtered - self.errors - self.data

        # Provide feedback to the user
        print(f" Read {headers} headers, {rows} lines. Filtered {self.filtered}. " +
              f"Errors {self.errors}. Bad data {self.data}. Wrote {loaded} records.")

    def parse_header(self, header: olieigra.HeaderModel) -> bool:
        """Transform the header record and start writing a record"""

        # Combine the separate fields into a date
        effective_date = datetime(header.year, header.month, header.day)

        # Skip the record if it is too old
        if effective_date < self.min_effective_date:
            self.filtered += 1
            return False

        # We may not write the row, so save the header columns in a variable for now
        self.hout = f'{header.id},{effective_date:%Y-%m-%d},{header.hour}'

        # Tell olieigra to process the body associated with this header
        return True
    
    def parse_body(self, body: list[olieigra.BodyModel]):
        """Perform some analytics and finish writing a record"""
        
        # Initialize
        usable_all = 0
        surface_nan = 1
        pres = []
        temp = []
        dp = []

        # Iterate through each record in the body
        for item in body:
            # We don't care about non-pressure records
            if item.type[0] == '3':
                continue

            # We don't want records that contain a NaN value
            if math.isnan(item.dpdp) | math.isnan(item.temp):
                continue

            # This is a usable record
            usable_all += 1

            # Flag that we found a valid surface record
            if item.type == '21':
                surface_nan = 0

            pres.append(item.pres / 100.0)
            temp.append(item.temp / 10.0)
            dp.append((item.temp - item.dpdp) / 10.0)

        # Quality checks
        if surface_nan == 1 or len(pres) < 20:
            self.data += 1
            return
        
        # Calulate lifted index
        li = self.calculate_li(pres, temp, dp)
        
        # If LI is NaN, there was an exception in the calulation
        if math.isnan(li):
            self.errors += 1
            return
        
        # Write record
        if surface_nan == 0:
            self.writer.write(f'{self.hout},{li:.1f}\n')

    def calculate_li(self, pres: list[float], temp: list[float], dp: list[float]) -> float:
        """Calculate the lifted index using metpy"""

        # Conver the units
        pres = np.array(pres) * units.hPa
        temp = np.array(temp) * units.degC
        dp = np.array(dp) * units.degC

        # Calculate parameters
        try:
            # Calculate the parcel profile
            prof = parcel_profile(pres, temp[0], dp[0])

            # The calculation throws an exception if the data doesn't reach the EL
            li = lifted_index(pres, temp, prof)

            return float(li.magnitude[0])
        except:
            return math.nan


# Set up for processing
callbacks = LiftedIndex(SILVER_LI_PATH, datetime(2000, 1, 1))
reader = olieigra.Reader(callbacks=callbacks)
crawler = olieigra.Crawler(reader=reader)

# Crawl and process files
crawler.crawl(BRONZE_DATA_POR_PATH)
