==> experiments/liftedindex_lr/12_calculate_features_labels.ipynb <==
# Create the Features and Labels Files
After calculating the lifted index, we will aggregate all the features and labels together and create one feature file and one label file.

import os
import glob
import pandas as pd

SILVER_GPH20S10K_PATH = '/usr/datalake/silver/igra/gph20s10k'
SILVER_LI_PATH = '/usr/datalake/silver/igra/li'
GOLD_PARQUET_PATH = '/usr/datalake/silver/igra/liftedindex_lr'

# You will need the pyarrow package to create a parquet file
#%conda install pyarrow

# Make sure the destiation folder exists
os.makedirs(GOLD_PARQUET_PATH, exist_ok=True)


def df_from_concatinated_csv(path: str, pattern: str) -> pd.DataFrame:
    """Load multiple csv files into one DataFrame"""
    
    # Get the list of files from a folder
    file_list = glob.glob(path + '/' + pattern)

    # Read each CSV file into DataFrame
    # This creates a list of dataframes
    df_list = (pd.read_csv(file) for file in file_list)

    # Pandas doesn't like to concatinate "empty" DataFrames
    df_list = [dataframe for dataframe in df_list if dataframe.shape[0] > 0]

    # Concatenate all DataFrames
    return pd.concat(df_list, ignore_index=True)

def df_merge(features: pd.DataFrame, labels: pd.DataFrame):
    """Perform inner join on raw features and labels"""

    # Inner Join on id, effectie_date, and hour
    return features \
        .merge(labels, how='inner', on=['id', 'effective_date', 'hour'])

def orchestrate() -> pd.DataFrame:
    """Load and merge the raw files"""
    df_labels = df_from_concatinated_csv(SILVER_LI_PATH, '*-data-li.csv')
    df_features = df_from_concatinated_csv(SILVER_GPH20S10K_PATH, '*-data-gph20s10k.csv')
    return df_merge(df_features, df_labels)


# Get a combined DataFrame of the raw feauters and labels
df = orchestrate()

# Sanity checks
print(f"Row count: {df.shape[0]}")
print("df.info:")
df.info()

# View statistics
df.describe()

# Write the parquet file.
df.to_parquet(GOLD_PARQUET_PATH + '/gph20s10k_li.parquet')
