==> experiments/liftedindex_lr/21_NotebookAutoencoderTrain.ipynb <==
import os
import pandas as pd
import pickle
import torch

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler

PARQUET_PATH = '/usr/datalake/silver/igra/liftedindex_lr/gph20s10k_li.parquet'
ARTIFACTS_PATH = '/usr/datalake/silver/igra/liftedindex_lr/artifacts'

batch_size = 256
epochs = 64
learning_rate = 0.001
learning_rate_gamma = 0.75

def load_dataset():
    X = pd.read_parquet(PARQUET_PATH)

    # Remove irrelevant data
    X = X.drop(['id', 'effective_date', 'hour', 'li'], axis=1)
    
    # Scale the X dataset
    ss = MinMaxScaler() 
    X = ss.fit_transform(X)

    # Save the transform
    os.makedirs(ARTIFACTS_PATH, exist_ok=True)
    with open(f'{ARTIFACTS_PATH}/ae_min_max_scaler.pkl', 'wb') as f:
        pickle.dump(ss, f)
    
    return train_test_split(X, test_size=0.2)

x_train, x_test = load_dataset()
print (f"Training size: {x_train.size:,}")
print (f"Predict size: {x_test.size:,}")
print (f"Feature count: {len(x_train[0])}")

x_train = torch.from_numpy(x_train).float().cuda()
x_test = torch.from_numpy(x_test).float().cuda()
n_batches = x_train.size()[0] // batch_size

class AutoEncoder(torch.nn.Module):
    def __init__(self):
        super().__init__()
        
        self.encoder = torch.nn.Sequential(
            torch.nn.Linear(127, 50),
            torch.nn.ReLU(),
            torch.nn.Linear(50, 10)
        )

        self.decoder = torch.nn.Sequential(
            torch.nn.Linear(10, 50),
            torch.nn.ReLU(),
            torch.nn.Linear(50, 127)
        )

    def forward(self, x):
        encoded = self.encoder(x)
        decoded = self.decoder(encoded)

        return decoded
    
model = AutoEncoder().cuda()
optimizer = torch.optim.Adam(model.parameters(), learning_rate)
loss_function = torch.nn.MSELoss()
scheduler = torch.optim.lr_scheduler.ExponentialLR(optimizer, gamma=learning_rate_gamma)

def train(model, loss_fn, optimizer, inputs, labels) -> float:
    optimizer.zero_grad()

    # Calculate error
    logits = model(inputs)
    cost = loss_fn(logits, labels)

    # Back propagation
    cost.backward()
    optimizer.step()

    return float(cost.item())

def predict(model, inputs):
    optimizer.zero_grad()

    # Calculate error
    logits = model(inputs)

    return logits

def r2_score_manual(preds, target):
    target_mean = torch.mean(target)
    ss_tot = torch.sum((target - target_mean) ** 2)  # Total sum of squares
    ss_res = torch.sum((target - preds) ** 2)      # Residual sum of squares
    r2 = 1 - (ss_res / ss_tot)
    return float(r2.item())

def train_orch():
    for epoch in range(epochs):
        cost = 0
        loader = torch.utils.data.DataLoader(dataset = x_train,
                                             batch_size = batch_size,
                                             shuffle = True)

        for batch in loader:
            cost += train(model, loss_function, optimizer, batch, batch)

        preds = predict(model, x_test)
        acc = r2_score_manual(x_test, preds)
        scheduler.step()

        print(f"Epoch: {epoch+1}, cost: {cost / n_batches:.4f}, acc: {acc:.3f}, lr: {scheduler.get_last_lr()[0]:.2e}")

train_orch()

torch.save(model.state_dict(), f'{ARTIFACTS_PATH}/ae_fnn.pt')
