==> experiments/liftedindex_lr/22_NotebookAutoencoderTest.ipynb <==
import pandas as pd
import pickle
import torch

from sklearn.preprocessing import MinMaxScaler

PARQUET_PATH = '/usr/datalake/silver/igra/liftedindex_lr/gph20s10k_li.parquet'
ARTIFACTS_PATH = '/usr/datalake/silver/igra/liftedindex_lr/artifacts'

def load_min_max_scaler()-> MinMaxScaler:
    with open(f'{ARTIFACTS_PATH}/ae_min_max_scaler.pkl', 'rb') as f:
        std_scaler = pickle.load(f)

    return std_scaler

class AutoEncoder(torch.nn.Module):
    def __init__(self):
        super().__init__()
        
        self.encoder = torch.nn.Sequential(
            torch.nn.Linear(127, 50),
            torch.nn.ReLU(),
            torch.nn.Linear(50, 10)
        )

        self.decoder = torch.nn.Sequential(
            torch.nn.Linear(10, 50),
            torch.nn.ReLU(),
            torch.nn.Linear(50, 127)
        )

    def forward(self, x):
        encoded = self.encoder(x)
        decoded = self.decoder(encoded)

        return decoded

# Load the dataset
df = pd.read_parquet(PARQUET_PATH)

# Grab 100 random samples
df = df.sample(100)
df = df.reset_index(drop=True)

# Remove irrelevant data
X = df.drop(['id', 'effective_date', 'hour', 'li'], axis=1)

# Scale the dataset
ss = load_min_max_scaler()
X = ss.transform(X)

# Convert to tensor
X = torch.from_numpy(X).float()
X.shape

model = AutoEncoder()
model.load_state_dict(torch.load(f'{ARTIFACTS_PATH}/ae_fnn.pt'))
model.eval()

# Encode the values
logits = model.encoder(X).detach()

# Decode the values
decoded = model.decoder(logits).detach()

# Undo the initial scaling transform
decoded = ss.inverse_transform(decoded).round(1)
decoded.shape

# Add the index back to the dataframe
output = pd.DataFrame(decoded, columns=df.columns[3:-1])
output.insert(0, 'id', df['id'] )
output.insert(1, 'effective_date', df['effective_date'] )
output.insert(2, 'hour', df['hour'] )
output.head()

# Interleave the results into the source for visual comparison
interleave = pd.concat([df, output])
interleave = interleave.sort_index()# .sort_values(by=['id', 'effective_date'])
interleave.head()

interleave.to_csv(f'{ARTIFACTS_PATH}/ae_compare.csv')
